refactor(settings-overlay): extract shared overlay contracts
Split SettingsOverlay.tsx into focused modules: types.ts (SettingFieldKey,
HelpFieldId, FieldHelp, SettingFieldConfig, SettingsSectionConfig and option
value types), constants.ts (ZOOM_STEP, ACCIDENTAL_OPTIONS,
ENHARMONIC_DISPLAY_OPTIONS, SETTING_FIELDS, SETTINGS_SECTIONS), shared.tsx
(OverlaySection, OverlayFieldHeader), useSettingsForm (seven atom-backed form
values), and useHelpPopover (activeHelpField, refs, toggle logic, document
mousedown listener). Reset confirmation stays in SettingsOverlaySurface.
Root useAtomValue subscriptions preserved in SettingsOverlay unchanged.
